Coding New performance model
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@349 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/isapi.cpp
@@ -648,8 +648,8 @@ BOOL buildAllRawHeaders(httpThreadContext* td,LPCONNECTION a,LPVOID output,LPDWO
 	{
 		valLen+=sprintf(&ValStr[valLen],"SCRIPT_NAME:");
 		lstrcpyn(&ValStr[valLen],td->request.URI,strlen(td->request.URI)-strlen(td->pathInfo)+1);
-		valLen+=strlen(td->request.URI)-strlen(td->pathInfo)+1;
-		valLen+=sprintf(&ValStr[valLen],"\n");
+		valLen+=(DWORD)strlen(td->request.URI)-strlen(td->pathInfo)+1;
+		valLen+=(DWORD)sprintf(&ValStr[valLen],"\n");
 	}
 	else if(valLen+30<maxLen) 
 		return 0;
